requirements.txt
flask
openai
pymysql
python-dotenv
gunicorn